pyproject.toml
[project]
name = "discord-bot"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "flask>=3.1.2",
    "python-dotenv>=1.2.1",
]
